requirements.txt
Django==4.2
Pillow==10.0.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.5.0
